+++ src/op/prop/add_edge_remove_edge.rs
@@ -24,7 +24,8 @@ where
     let mut clone = graph.clone();
 
     clone.add_edge(s, t);
-    clone.remove_edge(s, t);
+
+    let _ = clone.remove_edge(s, t);
 
     *graph == clone
 }